Refactor(plugins): Move jinja filter code for arista.avd.generate_lacp_id to PyAVD
