+++ python-avd/pyavd/templater.py
@@ -77,6 +77,7 @@ class Templar:
         from .j2filters.hide_passwords import hide_passwords
         from .j2filters.is_in_filter import is_in_filter
         from .j2filters.list_compress import list_compress
+        from .j2filters.markdown_rendering import status_render
         from .j2filters.natural_sort import natural_sort
         from .j2filters.snmp_hash import snmp_hash
         from .j2tests.contains import contains
@@ -101,6 +102,7 @@ class Templar:
                 "arista.avd.natural_sort": natural_sort,
                 "arista.avd.range_expand": range_expand,
                 "arista.avd.snmp_hash": snmp_hash,
+                "arista.avd.status_render": status_render,
             }
         )
         self.environment.tests.update(
